Moved daily btc tweet pull to tasks. Limited to 5k for demonstration purposes.
