+++ templates/404.html
@@ -1,4 +1,1 @@
-{% extends "/layout/page.html" %}
-{% block content %}
-    <img src="/static/img/404.jpg", width="100%">
-{% endblock %}
+
